FEAT: household profile members rendering
